+++ packages/config/tests/parse/tests.js
@@ -21,21 +21,53 @@ test('Configuration file - parsing error', async (t) => {
   await runFixture(t, 'parse_error')
 })
 
-test('Configuration file - ambiguous backslash in TOML', async (t) => {
-  await runFixture(t, 'parse_backslash_ambiguous', { flags: { featureFlags: { netlify_config_toml_backslash: true } } })
-})
-
-test('Configuration file - ambiguous backslash in TOML without feature flag', async (t) => {
-  await runFixture(t, 'parse_backslash_ambiguous', {
-    flags: { featureFlags: { netlify_config_toml_backslash: false } },
-  })
-})
-
-test('Configuration file - valid backslash in TOML', async (t) => {
+test('Configuration file - valid backslash sequences in TOML should not warn', async (t) => {
   const { returnValue } = await runFixture(t, 'parse_backslash_valid', {
     flags: { featureFlags: { netlify_config_toml_backslash: true } },
   })
   t.true(returnValue.includes('\\\\[this\\\\]\\ntest \\" \\b \\t \\n \\f \\r \\u0000 \\u0000'))
+})
+
+test('Configuration file - invalid backslash sequences in TOML without feature flag should not warn', async (t) => {
+  await runFixture(t, 'parse_backslash_double', {
+    flags: { featureFlags: { netlify_config_toml_backslash: false } },
+  })
+})
+
+test('Configuration file - invalid backslash sequences in double quote strings in TOML should warn', async (t) => {
+  await runFixture(t, 'parse_backslash_double', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
+})
+
+test('Configuration file - trailing backslashes in double quote strings in TOML should warn', async (t) => {
+  await runFixture(t, 'parse_backslash_trailing_double', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
+})
+
+test('Configuration file - invalid backslash sequences in multiline double quote strings in TOML should warn', async (t) => {
+  await runFixture(t, 'parse_backslash_double_multiline', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
+})
+
+test('Configuration file - trailing backslashes in multiline double quotes strings in TOML should not warn', async (t) => {
+  await runFixture(t, 'parse_backslash_trailing_double_multiline', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
+})
+
+test('Configuration file - invalid backslash sequences in single quote strings in TOML should not warn', async (t) => {
+  await runFixture(t, 'parse_backslash_single', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
+})
+
+test('Configuration file - invalid backslash sequences in multiline single quote strings in TOML should not warn', async (t) => {
+  await runFixture(t, 'parse_backslash_single_multiline', {
+    flags: { featureFlags: { netlify_config_toml_backslash: true } },
+  })
 })
 
 test('Redirects - redirects file', async (t) => {