+++ packages/functions-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### [2.1.7](https://www.github.com/netlify/build/compare/functions-utils-v2.1.6...functions-utils-v2.1.7) (2021-10-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to ^4.23.6 ([#3699](https://www.github.com/netlify/build/issues/3699)) ([ebf3385](https://www.github.com/netlify/build/commit/ebf33850f60f318ffd137c9838d3d3139c8abe28))
+
 ### [2.1.6](https://www.github.com/netlify/build/compare/functions-utils-v2.1.5...functions-utils-v2.1.6) (2021-10-06)